+++ rollup.congig.browser.js
@@ -1,0 +1,6 @@
+export default {
+    format: 'umd',
+    moduleName: 'li-log',
+    entry: 'src/browser.js',
+    dest: 'li-log.browser.js'
+};