+++ src/browser-version/console.js
@@ -1,0 +1,7 @@
+'use strict';
+
+import { Log } from './log.browser.js';
+
+export function toConsole(options) {
+    return new Log(options);
+}