+++ src/browser.js
@@ -1,0 +1,7 @@
+import { toConsole as console} from './browser-version/console';
+import { Log } from './browser-version/log.browser';
+
+Log.console = console;
+
+export default Log;
+