+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2017-09-15
+### Added
+- Base rollup config
+- Simple console output
+
 ## [0.1.0] - 2017-09-12
 ### Added
 - Project structure